server/requirements.txt
fastapi
passlib
sqlalchemy
injector 
elasticsearch
transformers
elasticsearch-dsl
langchain-groq